plugin WES :
* change CGX files for WES v1 and v2 with last firmware
* check the firmware revision at the startup of the plugin. The firmware revision should be greater than V0.80
* add the third TIC counter
* fix initialization bug with WES v1
* remove analog values for WES v1
* fix conversion for pulse value
* fix conversion for tele-info counters
* fix conversion for analog value

+++ sources/plugins/WES/equipments/WESEquipment.cpp
@@ -486,7 +486,7 @@ namespace equipments
          {
             try
             {
-               m_AnalogList[counter]->set(results.get<unsigned int>("ad" + boost::lexical_cast<std::string>(counter + 1)));
+               m_AnalogList[counter]->set((unsigned int)results.get<float>("ad" + boost::lexical_cast<std::string>(counter + 1)));
                keywordsToHistorize.push_back(m_AnalogList[counter]);
             }
             catch (std::exception& e)
